Add files via upload
Tiriso ya mafoko application is basically a simple game of matching verbs to nouns in relation to their use, for example eating will match food. The app was built using python and kivy, so it is recommend you update your graphic's driver. The words that are matched are in Setswana language, it was initially made for Setswana speakers from Botswana, however the language can be changed by replacing the text in Tiriso_ya_mafoko.txt with the text of language of your choice.